+++ packages/create-analog/files/analog-env.d.ts
@@ -1,0 +1,13 @@
+declare module '*.analog' {
+  import { Type } from '@angular/core';
+
+  const cmp: Type<unknown>;
+  export default cmp;
+}
+
+declare module '*.ag' {
+  import { Type } from '@angular/core';
+
+  const cmp: Type<unknown>;
+  export default cmp;
+}