+++ packages/create-analog/files/tailwind.config.cjs
@@ -1,8 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ['./index.html', './src/**/*.{html,ts,md}'],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};