+++ packages/create-analog/files/tailwind.config.ts
@@ -1,0 +1,9 @@
+import type { Config } from 'tailwindcss';
+
+export default {
+  content: ['./index.html', './src/**/*.{html,ts,md,analog,ag}'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+} satisfies Config;