+++ packages/create-analog/template-latest/vite.config.ts
@@ -11,7 +11,7 @@ export default defineConfig(({ mode }) => ({
   resolve: {
     mainFields: ['module'],
   },
-  plugins: [analog()],
+  plugins: [analog(__ANALOG_SFC_CONFIG__)],
   test: {
     globals: true,
     environment: 'jsdom',