+++ packages/platform/src/lib/options.ts
@@ -53,6 +53,12 @@ export interface Options {
    * Additional files to include in compilation
    */
   include?: string[];
+  /**
+   * Toggles internal API middleware.
+   * If disabled, a proxy request is used to route /api
+   * requests to / in the production server build.
+   */
+  useAPIMiddleware?: boolean;
 }
 
 export interface PrerenderContentDir {