+++ packages/vite-plugin-nitro/src/lib/options.ts
@@ -19,6 +19,14 @@ export interface Options {
    * Additional API paths to include
    */
   additionalAPIDirs?: string[];
+  apiPrefix?: string;
+
+  /**
+   * Toggles internal API middleware.
+   * If disabled, a proxy request is used to route /api
+   * requests to / in the production server build.
+   */
+  useAPIMiddleware?: boolean;
 }
 
 export interface PrerenderOptions {