+++ packages/vite-plugin-nitro/src/lib/vite-nitro-plugin.spec.data.ts
@@ -36,6 +36,13 @@ export const mockNitroConfig: NitroConfig = {
       },
     ],
   },
+  routeRules: undefined,
+  runtimeConfig: {
+    apiPrefix: 'api',
+  },
+  virtual: {
+    '#ANALOG_API_MIDDLEWARE': expect.anything(),
+  },
 };
 
 export async function mockBuildFunctions() {